Reboot
The old design had too many concepts to be easily explained.
This is a redesign and reimplementation with fewer concepts, but a
nearly identical CLI.

This redesign was done in three phases: initial concept design,
reimplementation and final concept documentation. Many details of the
design shifted throughout each of the phases, and little effort has been
put into cleaning up the mess.

Consequently the implementation is somewhat low quality, but the
concept documentation feels quite good, IMHO. However the effect handler
section is still quite thin.